Fixing lint errors
- Running with type-check on circleci
