+++ src/app/app.module.ts
@@ -22,7 +22,8 @@ import { AngularFireDatabaseModule } from 'angularfire2/database';
 import { AngularFireAuthModule } from 'angularfire2/auth';
 
 import { AppComponent } from './app.component';
-import { HomeComponent, AddCourseDialogComponent } from './home/home.component';
+import { HomeComponent } from './home/home.component';
+import { AddCourseDialogComponent } from './home/add-course-dialog.component';
 import { environment } from '../environments/environment';
 import { LoginComponent } from './login/login.component';
 import { AuthGuard } from './auth.guard';
